widen loop counter to unsigned int, to quieten checkers https://github.com/gphoto/libgphoto2/security/code-scanning/189

+++ camlibs/sonydscf55/sony.c
@@ -302,7 +302,7 @@ static int sony_packet_read(Camera * camera, Packet * pack)
 static int
 sony_packet_write(Camera * camera, Packet * p)
 {
-	unsigned short int count;
+	unsigned int count;
 	int rc;
 
 	GP_DEBUG(
